Move the code to `parseExtraField` into a separate file and use it everywhere we look at `extraField` (i.e. `Open/directory.js`, `Open/unzip.js` and `parse.js`)
Also fix test to look at headers and confirm the reported size is correct
