+++ lib/Open/directory.js
@@ -3,6 +3,7 @@ var PullStream = require('../PullStream');
 var unzip = require('./unzip');
 var Promise = require('bluebird');
 var BufferStream = require('../BufferStream');
+var parseExtraField = require('../parseExtraField');
 
 var signature = Buffer(4);
 signature.writeUInt32LE(0x06054b50,0);
@@ -62,6 +63,7 @@ module.exports = function centralDirectory(source) {
           return records.pull(vars.extraFieldLength);
         })
         .then(function(extraField) {
+          vars.extra = parseExtraField(extraField, vars);
           return records.pull(vars.fileCommentLength);
         })
         .then(function(comment) {
