+++ lib/Open/unzip.js
@@ -4,6 +4,7 @@ var PullStream = require('../PullStream');
 var Stream = require('stream');
 var binary = require('binary');
 var zlib = require('zlib');
+var parseExtraField = require('../parseExtraField');
 
 // Backwards compatibility for node versions < 8
 if (!Stream.Writable || !Stream.Writable.prototype.destroy)
@@ -40,35 +41,8 @@ module.exports = function unzip(source,offset,_password) {
           return file.pull(vars.extraFieldLength);
         })
         .then(function(extraField) {
-          var extra, checkEncryption;
-          // Find the ZIP64 header, if present.
-          while(!extra && extraField && extraField.length) {
-            var candidateExtra = binary.parse(extraField)
-              .word16lu('signature')
-              .word16lu('partsize')
-              .word64lu('uncompressedSize')
-              .word64lu('compressedSize')
-              .word64lu('offset')
-              .word64lu('disknum')
-              .vars;
-
-            if(candidateExtra.signature === 0x0001) {
-              extra = candidateExtra;
-            } else {
-              // Advance the buffer to the next part.
-              // The total size of this part is the 4 byte header + partsize.
-              extraField = extraField.slice(candidateExtra.partsize + 4);
-            }
-          }
-
-          extra = extra || {};
-
-          if (vars.compressedSize === 0xffffffff)
-            vars.compressedSize = extra.compressedSize;
-
-          if (vars.uncompressedSize  === 0xffffffff)
-            vars.uncompressedSize= extra.uncompressedSize;
-
+          var checkEncryption;
+          vars.extra = parseExtraField(extraField, vars);
           if (vars.flags & 0x01) checkEncryption = file.pull(12)
             .then(function(header) {
               if (!_password)
