+++ lib/parse.js
@@ -6,6 +6,7 @@ var Promise = require('bluebird');
 var PullStream = require('./PullStream');
 var NoopStream = require('./NoopStream');
 var BufferStream = require('./BufferStream');
+var parseExtraField = require('./parseExtraField');
 
 // Backwards compatibility for node versions < 8
 if (!Stream.Writable || !Stream.Writable.prototype.destroy)
@@ -108,20 +109,7 @@ Parse.prototype._readFile = function () {
       }
        
       self.pull(vars.extraFieldLength).then(function(extraField) {
-        var extra = binary.parse(extraField)
-          .word16lu('signature')
-          .word16lu('partsize')
-          .word64lu('uncompressedSize')
-          .word64lu('compressedSize')
-          .word64lu('offset')
-          .word64lu('disknum')
-          .vars;
-       
-        if (vars.compressedSize === 0xffffffff)
-          vars.compressedSize = extra.compressedSize;
-        
-        if (vars.uncompressedSize  === 0xffffffff)
-          vars.uncompressedSize= extra.uncompressedSize;
+        var extra = parseExtraField(extraField, vars);
 
         entry.vars = vars;
         entry.extra = extra;
