+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "unzipper",
-  "version": "0.9.0",
+  "version": "0.9.1",
   "description": "Unzip cross-platform streaming API ",
   "author": "Evan Oxfeld <eoxfeld@gmail.com>",
   "contributors": [
